Fix Pagination DoS Vulnerability
Explicitly cast the `page` input parameter to an integer in `ContatoController@index` to prevent a TypeError when an array is passed as the page parameter. Also added a test case.

+++ app/Http/Controllers/ContatoController.php
@@ -25,7 +25,7 @@ class ContatoController extends Controller
         $contatos = Contato::latest()->paginate(5);
 
         return view('contato.index', compact('contatos'))
-            ->with('i', (request()->input('page', 1) - 1) * 5);
+            ->with('i', ((int) request()->input('page', 1) - 1) * 5);
     }
 
     /**
